Move use-sticky.ts to src/client/side-effects
https://youtrack.weseek.co.jp/issue/GW-7946
- Move use-sticky.ts to src/client/side-effects and update imports
