+++ packages/app/src/components/Fab.tsx
@@ -5,10 +5,10 @@ import React, {
 import { animateScroll } from 'react-scroll';
 import { useRipple } from 'react-use-ripple';
 
+import { useSticky } from '~/client/side-effects/use-sticky';
 import { usePageCreateModal } from '~/stores/modal';
 import { useCurrentPagePath } from '~/stores/page';
 import { useIsAbleToChangeEditorMode } from '~/stores/ui';
-import { useSticky } from '~/stores/use-sticky';
 import loggerFactory from '~/utils/logger';
 
 import { CreatePageIcon } from './Icons/CreatePageIcon';
